Annotate case node DFA transitions with source DFA transition data
Also, instead of just reporting which DFAs are in conflict, trace as
precisely as possible which transitions and/or states cause the
ambiguity. This allows the error formatter to point at (e.g.) specific
parts of a bad regex in the best case.

(tracking transitions into states is necessary as all the existing
matches funnel everything into a single accept state if at all possible)